refactor: update imports to use shadcn directory in dashboard pages
